backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
psycopg2==2.9.9 
python-dateutil==2.8.2
feedparser==6.0.11
openai==1.59.3
apscheduler==3.10.4

